Initial Duel app files
