+++ api/.env
@@ -1,0 +1,9 @@
+#default to 3000
+PORT=3000
+
+# Database credentials
+DB_HOST='localhost'
+DB_USER='postgres'
+DB_PASS='123456'
+DB_NAME='MASS'
+DB_PORT=5432
