+++ api/README.md
@@ -1,4 +1,4 @@
-# RateMyCourses
+# RateMyCourses API Server
 
 ## How to start
 If first time then download all dependencies with:
@@ -7,6 +7,6 @@ npm install
 ```
 Start server with:
 ```
-node app.js
+npm start
 ```
