+++ api/config.js
@@ -1,0 +1,14 @@
+require('dotenv').config();
+
+const pgp = require('pg-promise')();
+
+const db = pgp({
+    host: process.env.DB_HOST,
+    user: process.env.DB_USER,
+    database: process.env.DB_NAME,
+    password: process.env.DB_PASS,
+    port: process.env.DB_PORT,
+});
+
+module.exports = db;
+
