+++ api/package.json
@@ -4,19 +4,22 @@
   "description": "Web App for finding courses",
   "private": true,
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node ./bin/www"
   },
   "license": "ISC",
   "dependencies": {
+    "@material-ui/core": "latest",
+    "cors": "^2.8.5",
     "dotenv": "^8.2.0",
-    "ejs": "~2.6.1",
     "express": "~4.16.1",
-    "express-react-views": "^0.11.0",
     "express-session": "^1.17.0",
     "express-validator": "^6.4.0",
     "jquery": "^3.4.1",
+    "morgan": "~1.9.1",
     "pg-promise": "10.4.3",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1"
+    "react": "latest",
+    "react-dom": "latest",
+    "react-scripts": "latest"
   }
 }
