--- config.js
@@ -1,13 +1,0 @@
-require('dotenv').config();
-
-module.exports = {
-    port: 8888,
-    db: {
-   	 host: 'localhost',
-   	 user: 'INSERT USER',
-  	 database: 'INSERT DATABASE NAME',
-  	 password: 'INSERT PASSWORD',
-  	 port: 5432,
-	}
-};
-
