--- public/css/global.css
@@ -1,128 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-}
-
-body {
-    background-color: #f5f5f5;
-}
-
-.button {
-    display: block;
-    -webkit-appearance: none;
-    outline: none;
-    border: none;
-    background-color: white;
-    cursor: pointer;
-    border-radius: 3px;
-    box-shadow: 0px 2px 4px -1px rgba(0,0,0,0.1), 0px 4px 5px 0px rgba(0,0,0,0.07), 0px 1px 10px 0px rgba(0,0,0,0.06);
-    font-family: Roboto;
-}
-
-.button--dark {
-
-}
-
-.button--rounded {
-    border-radius: 50%;
-}
-
-.header-container {
-    width: 100vw;
-    height: 100px;
-    background-color: #747ce0;
-    display: flex;
-    box-shadow: 0px 2px 4px -1px rgba(0,0,0,0.2), 0px 4px 5px 0px rgba(0,0,0,0.14), 0px 1px 10px 0px rgba(0,0,0,0.12);
-    position: relative;
-}
-
-.header-title {
-    display: flex;
-    color: white;
-    letter-spacing: 4px;
-    font-family: Raleway;
-    font-size: 54px;
-    text-transform: uppercase;
-}
-
-.header-left-container {
-    display: flex;
-    height: 100%;
-    align-items: center;
-    left: 24px;
-    position: absolute;
-}
-
-.header-middle-container {
-    display: flex;
-    height: 100%;
-    align-items: center;
-    margin-left: auto;
-    margin-right: auto;
-}
-
-.header-right-container {
-    display: flex;
-    height: 100%;
-    align-items: center;
-    right: 24px;
-    position: absolute;
-}
-
-.header-search-container {
-    position: relative;
-}
-
-.header-search-icon {
-    display: block;
-    position: absolute;
-    left: 12px;
-    top: 10px;
-    width: 28px;
-    height: auto;
-    pointer-events: none;
-}
-
-.header-search-input {
-    display: block;
-    height: 46px;
-    width: 400px;
-    outline: none;
-    font-size: 24px;
-    border-radius: 23px;
-    border: none;
-    padding-left: 50px;
-    padding-right: 20px;
-    box-sizing: border-box;
-    text-overflow: ellipsis;
-    box-shadow: 0px 2px 4px -1px rgba(0,0,0,0.2), 0px 4px 5px 0px rgba(0,0,0,0.14), 0px 1px 10px 0px rgba(0,0,0,0.12);
-}
-
-.header-button {
-    margin-left: 22px;
-    padding: 12px 0px;
-    width: 120px;
-    font-size: 18px;
-    text-transform: uppercase;
-}
-
-.header-icon-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    cursor: pointer;
-}
-
-.header-catalog-icon,
-.header-heart-icon,
-.header-user-icon {
-    width: 52px;
-    height: auto;
-    fill: white;
-}
-
-.header-heart-icon,
-.header-user-icon {
-    margin-left: 19px;
-
-}
