--- views/pages/components/header.jsx
@@ -1,60 +1,0 @@
-const React = require('react');
-
-const SearchIcon = () => (
-    <svg class="header-search-icon" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="black" width="18px" height="18px"><path d="M0 0h24v24H0z" fill="none"/><path d="M15.5 14h-.79l-.28-.27C15.41 12.59 16 11.11 16 9.5 16 5.91 13.09 3 9.5 3S3 5.91 3 9.5 5.91 16 9.5 16c1.61 0 3.09-.59 4.23-1.57l.27.28v.79l5 4.99L20.49 19l-4.99-5zm-6 0C7.01 14 5 11.99 5 9.5S7.01 5 9.5 5 14 7.01 14 9.5 11.99 14 9.5 14z"/></svg>
-)
-
-const CatalogIcon = () => (
-    <svg class="header-catalog-icon" xmlns="http://www.w3.org/2000/svg" enable-background="new 0 0 24 24" viewBox="0 0 24 24" fill="black" width="18px" height="18px"><g><rect fill="none" height="24" width="24"/></g><g><g><g><path d="M17.5,4.5c-1.95,0-4.05,0.4-5.5,1.5c-1.45-1.1-3.55-1.5-5.5-1.5S2.45,4.9,1,6v14.65c0,0.65,0.73,0.45,0.75,0.45 C3.1,20.45,5.05,20,6.5,20c1.95,0,4.05,0.4,5.5,1.5c1.35-0.85,3.8-1.5,5.5-1.5c1.65,0,3.35,0.3,4.75,1.05 C22.66,21.26,23,20.86,23,20.6V6C21.51,4.88,19.37,4.5,17.5,4.5z M21,18.5c-1.1-0.35-2.3-0.5-3.5-0.5c-1.7,0-4.15,0.65-5.5,1.5V8 c1.35-0.85,3.8-1.5,5.5-1.5c1.2,0,2.4,0.15,3.5,0.5V18.5z"/></g></g></g></svg>
-);
-
-const HeartIcon = () => (
-    <svg class="header-heart-icon" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="black" width="18px" height="18px"><path d="M0 0h24v24H0z" fill="none"/><path d="M12 21.35l-1.45-1.32C5.4 15.36 2 12.28 2 8.5 2 5.42 4.42 3 7.5 3c1.74 0 3.41.81 4.5 2.09C13.09 3.81 14.76 3 16.5 3 19.58 3 22 5.42 22 8.5c0 3.78-3.4 6.86-8.55 11.54L12 21.35z"/></svg>
-)
-
-const UserIcon = () => (
-    <svg class="header-user-icon" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="black" width="18px" height="18px"><path d="M0 0h24v24H0z" fill="none"/><path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm0 3c1.66 0 3 1.34 3 3s-1.34 3-3 3-3-1.34-3-3 1.34-3 3-3zm0 14.2c-2.5 0-4.71-1.28-6-3.22.03-1.99 4-3.08 6-3.08 1.99 0 5.97 1.09 6 3.08-1.29 1.94-3.5 3.22-6 3.22z"/></svg>
-)
-
-const Header = (props) => {
-	return (
-        <div className='header-container'> 
-            <div className='header-left-container'>
-                <span className="header-title">Rate My Courses</span>
-            </div>
-            {
-                !props.hideSearch && (
-                    <div className='header-middle-container'>
-                        <div className="header-search-container">
-                            <SearchIcon/>
-                            <input type="text" className="header-search-input"></input>
-                        </div>
-                    </div>
-                )
-            }
-            <div className='header-right-container'>
-                <div className="header-icon-container">
-                    <CatalogIcon/>
-                </div>
-                { props.user ? (
-                    <React.Fragment>
-                        <div className="header-icon-container">
-                            <HeartIcon/>
-                        </div>
-                        <div className="header-icon-container">
-                            <UserIcon/>
-                        </div>
-                        <input type="button" className="header-button button" value="Logout"></input>
-                    </React.Fragment>
-                ) : (
-                    <React.Fragment>
-                        <input type="button" className="header-button button" value="Login"></input>
-                        <input type="button" className="header-button button" value="Register"></input>
-                    </React.Fragment>
-                )}
-            </div>
-        </div>
-	);
-}
-
-module.exports = Header;
